% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopword.R
\name{stopword}
\alias{stopword}
\title{StopWords do Thesa}
\usage{
stopword(l)
}
\arguments{
\item{l}{linguage do stopword}
}
\value{
Retorna as StopWords
}
\description{
Recupera as stopswords do Thesa
}
\details{
Utilize este campo para escrever detalhes mais tecnicos da
sua funcao (se necessario), ou para detalhar melhor como
utilizar determinados argumentos.
}
\examples{
stopword("pt")

x <- 3
y <- 4
soma(x = x, y = y)

}
\seealso{
\code{\link[base]{sum}}, \code{\link[base]{+}}
}
\author{
Rene Faustino Gabriel Junior
}
